Corrected Beads Calculation and Calibration Round
No 100% what was wrong with beads calculation, but after some debugging it seems it is calculating the spectra all right. Although, peaks are a little off from the nominal peak value. Probably the code was selecting the fitted gaussian peaks for the 680nm channel and this added a lot of variation to the image mask and this jiggled the peaks of the red and green channels.